revert to previous legacy order + gather new module computation classes in a dedicated folder
